tests: move common tooling to separate file
Reorganize the tests a little bit so that we can more easily share
common functions. We could keep all this code in the tests directory but
it seems a little cleaner to move it to src.

Also, rename the test file that we currently have from integration
to get_resource as it will contain tests based around getting resources.
