+++ integration-tests/src/lib.rs
@@ -1,0 +1,5 @@
+// Copyright (c) 2025 by IBM.
+// Licensed under the Apache License, Version 2.0, see LICENSE for details.
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod common;
